Ensure to reference the before/after WixAction when scheduling an action
Fixes 7225
