+++ src/wix/test/WixToolsetTest.CoreIntegration/MsiFixture.cs
@@ -520,38 +520,6 @@ namespace WixToolsetTest.CoreIntegration
         }
 
         [Fact]
-        public void CanBuildSetProperty()
-        {
-            var folder = TestData.Get(@"TestData\SetProperty");
-
-            using (var fs = new DisposableFileSystem())
-            {
-                var baseFolder = fs.GetFolder();
-                var intermediateFolder = Path.Combine(baseFolder, "obj");
-
-                var result = WixRunner.Execute(new[]
-                {
-                    "build",
-                    Path.Combine(folder, "Package.wxs"),
-                    Path.Combine(folder, "PackageComponents.wxs"),
-                    "-loc", Path.Combine(folder, "Package.en-us.wxl"),
-                    "-bindpath", Path.Combine(folder, "data"),
-                    "-intermediateFolder", intermediateFolder,
-                    "-o", Path.Combine(baseFolder, @"bin\test.msi")
-                });
-
-                result.AssertSuccess();
-
-                var output = WindowsInstallerData.Load(Path.Combine(baseFolder, @"bin\test.wixpdb"), false);
-                var caRows = output.Tables["CustomAction"].Rows.Single();
-                WixAssert.StringEqual("SetINSTALLLOCATION", caRows.FieldAsString(0));
-                WixAssert.StringEqual("51", caRows.FieldAsString(1));
-                WixAssert.StringEqual("INSTALLLOCATION", caRows.FieldAsString(2));
-                WixAssert.StringEqual("[INSTALLFOLDER]", caRows.FieldAsString(3));
-            }
-        }
-
-        [Fact]
         public void CanBuildVersionIndependentProgId()
         {
             var folder = TestData.Get(@"TestData\ProgId");
